code formmating / show options available in array
