+++ values.php
@@ -1,0 +1,45 @@
+<?php
+// values of different lengths to 1 meter
+const meter = array(
+    "Inch" => 0.0254,
+    "Foot" => 0.3048,
+    "Yard" => 0.9144,
+    "Mile" => 1609.344,
+    "Millimeter" => 0.001,
+    "Centimeter" => 0.01,
+    "Meter" => 1,
+    "Kilometer" => 1000,
+    "Micrometer" => 0.000001,
+    "Nanometer" => 0.000000001
+  );
+  
+  // values of different volumes to 1 liter
+  const liter = array(
+    "Cubic Meter" => 0.001,
+    "Cubic Centimeter" => 1000,
+    "Cubic Millimeter" => 1,
+    "Liter" => 0.000001
+  );  
+  
+  // values of different weight to 1 kilogram
+  const kg = array(
+    "Ounce" =>	0.0283495,
+    "Pound" =>	0.453592,
+    "stones" =>	6.35029,
+    "Tonne" =>	1000,
+    "Gram" =>	0.001,
+    "Kilogram" =>	1
+  );
+  
+  // values of different speeds to 1 KMPH
+  const kmph = array(
+      "Lightspeed" => 1079252848.8,
+      "Mach" => 1255.08,
+      "Mlph" => 1.609344,
+      "Mps" => 3.6,
+      "Kmph" => 1,
+      "Kmps" => 3600,
+      "Knot" => 1.852
+  );
+
+?>
