+++ weight.php
@@ -1,3 +1,6 @@
+<?php
+require "values.php";
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -28,11 +31,11 @@
               <div class="form-group">
                 <label for="opt1">From</label>
                 <select name="opt1" id="opt1" class="form-control">
-                  <option value="Kilogram">Kilo Gram</option>
-                  <option value="Gram">Gram</option>
-                  <option value="Tonne">Tonne</option>
-                  <option value="Pound">Pound</option>
-                  <option value="Ounce">Ounce</option>
+                  <?php
+                    foreach (array_keys(kg) as $unit) {
+                      echo "<option value='$unit'>$unit</option>";
+                    }
+                  ?>
                 </select>
               </div>
               <div class="form-group">
@@ -41,11 +44,11 @@
               <div class="form-group">
                 <label for="opt2">To</label>
                 <select name="opt2" id="opt2" class="form-control">
-                  <option value="Kilogram">Kilo Gram</option>
-                  <option value="Gram">Gram</option>
-                  <option value="Tonne">Tonne</option>
-                  <option value="Pound">Pound</option>
-                  <option value="Ounce">Ounce</option>
+                  <?php
+                    foreach (array_keys(kg) as $unit) {
+                      echo "<option value='$unit'>$unit</option>";
+                    }
+                  ?>
                 </select>
               </div>
               <div class="form-group">
